+++ packages/postgres-database/src/simcore_postgres_database/utils_aiopg_orm.py
@@ -6,6 +6,8 @@
         - the new async sqlalchemy ORM https://docs.sqlalchemy.org/en/14/orm/
         - https://piccolo-orm.readthedocs.io/en/latest/index.html
 """
+# pylint: disable=no-value-for-parameter
+
 import functools
 import operator
 from typing import Any, Generic, TypeVar
@@ -16,8 +18,6 @@ from aiopg.sa.result import ResultProxy, RowProxy
 from sqlalchemy import func
 from sqlalchemy.sql.base import ImmutableColumnCollection
 from sqlalchemy.sql.dml import Insert, Update, UpdateBase
-
-# pylint: disable=no-value-for-parameter
 from sqlalchemy.sql.elements import literal_column
 from sqlalchemy.sql.schema import Column
 from sqlalchemy.sql.selectable import Select
